Move ScriptMsg from msg crate into script_traits
