+++ components/layout_traits/lib.rs
@@ -24,10 +24,10 @@ extern crate util;
 use gfx::font_cache_task::FontCacheTask;
 use gfx::paint_task::LayoutToPaintMsg;
 use ipc_channel::ipc::{IpcReceiver, IpcSender};
-use msg::constellation_msg::ScriptMsg as ConstellationMsg;
 use msg::constellation_msg::{ConstellationChan, Failure, PipelineId};
 use net_traits::image_cache_task::ImageCacheTask;
 use profile_traits::{mem, time};
+use script_traits::ScriptMsg as ConstellationMsg;
 use script_traits::{LayoutControlMsg, ConstellationControlMsg, OpaqueScriptLayoutChannel};
 use url::Url;
 use util::ipc::OptionalIpcSender;
