+++ components/script/dom/canvasrenderingcontext2d.rs
@@ -35,10 +35,10 @@ use euclid::point::Point2D;
 use euclid::rect::Rect;
 use euclid::size::Size2D;
 use ipc_channel::ipc::{self, IpcSender};
-use msg::constellation_msg::ScriptMsg as ConstellationMsg;
 use net_traits::image::base::PixelFormat;
 use net_traits::image_cache_task::ImageResponse;
 use num::{Float, ToPrimitive};
+use script_traits::ScriptMsg as ConstellationMsg;
 use std::str::FromStr;
 use std::sync::mpsc::channel;
 use std::{cmp, fmt};
