+++ components/script/dom/htmliframeelement.rs
@@ -25,10 +25,10 @@ use dom::window::Window;
 use js::jsapi::{JSAutoCompartment, JSAutoRequest, RootedValue, JSContext, MutableHandleValue};
 use js::jsval::{UndefinedValue, NullValue};
 use msg::constellation_msg::IFrameSandboxState::{IFrameSandboxed, IFrameUnsandboxed};
-use msg::constellation_msg::ScriptMsg as ConstellationMsg;
 use msg::constellation_msg::{ConstellationChan, IframeLoadInfo, MozBrowserEvent};
 use msg::constellation_msg::{NavigationDirection, PipelineId, SubpageId};
 use page::IterablePage;
+use script_traits::ScriptMsg as ConstellationMsg;
 use std::ascii::AsciiExt;
 use std::cell::Cell;
 use string_cache::Atom;
