+++ components/script/dom/htmltextareaelement.rs
@@ -25,9 +25,9 @@ use dom::node::{document_from_node, window_from_node};
 use dom::nodelist::NodeList;
 use dom::virtualmethods::VirtualMethods;
 use msg::constellation_msg::ConstellationChan;
-use msg::constellation_msg::ScriptMsg as ConstellationMsg;
 use script_task::ScriptTaskEventCategory::InputEvent;
 use script_task::{CommonScriptMsg, Runnable};
+use script_traits::ScriptMsg as ConstellationMsg;
 use selectors::states::*;
 use std::cell::Cell;
 use string_cache::Atom;
