+++ components/script/layout_interface.rs
@@ -14,13 +14,12 @@ use ipc_channel::ipc::{IpcReceiver, IpcSender};
 use libc::uintptr_t;
 use msg::compositor_msg::Epoch;
 use msg::compositor_msg::LayerId;
-use msg::constellation_msg::ScriptMsg as ConstellationMsg;
 use msg::constellation_msg::{ConstellationChan, Failure, PipelineId};
 use msg::constellation_msg::{WindowSizeData};
 use net_traits::image_cache_task::ImageCacheTask;
 use profile_traits::mem::ReportsChan;
-use script_traits::{ConstellationControlMsg, LayoutControlMsg};
-use script_traits::{OpaqueScriptLayoutChannel, UntrustedNodeAddress};
+use script_traits::{ConstellationControlMsg, LayoutControlMsg, OpaqueScriptLayoutChannel};
+use script_traits::{ScriptMsg as ConstellationMsg, UntrustedNodeAddress};
 use selectors::parser::PseudoElement;
 use std::any::Any;
 use std::sync::Arc;
